Fix background script losing connection to devtools page
When the background script is suspended after a while of inactivity, it
loses all its state and thus the connection port to the open devpage.

When we open the popup again, the background page is reloaded, but the
devtools connection isn't. Now we ask any open devtools window for a new
port whenever the popup is opened and we don't have any devtool ports
available.

Fixes #3